pyproject.toml
[project]
name = "used-car-price-prediction"
version = "0.0.1"
description = "This is a simple project which applies machine learning to predict price of used car and help users pay the right price for the used car they decide to buy."
authors = [{ name = "minhquana1906", email = "quann1906@gmail.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.12,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "fastapi[standard]>=0.115.11",
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "isort>=6.0.1",
    "joblib>=1.4.2",
    "loguru>=0.7.3",
    "matplotlib>=3.9.4",
    "numpy==1.26.4",
    "pandas>=2.2.3",
    "pyodbc>=5.2.0",
    "scikit-learn==1.2.2",
    "seaborn>=0.13.2",
    "streamlit>=1.43.2",
    "tqdm>=4.67.1",
    "xgboost==2.0.3",
]

[project.urls]
Homepage = "https://minhquana1906.github.io/used-car-price-prediction/"
Repository = "https://github.com/minhquana1906/used-car-price-prediction"
Documentation = "https://minhquana1906.github.io/used-car-price-prediction/"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    "deptry>=0.22.0",
    "mypy>=0.991",
    "ruff>=0.9.2",
    "category-encoders>=2.6.4",
    "optuna>=4.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.setuptools]
py-modules = ["used_car_price_prediction"]

[tool.mypy]
files = ["used_car_price_prediction"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true
